Fix false positive javascript error "w2wDataTables: Ajax call returned a undefined status code"
When a XHR request is triggered while a former request is still pending, this former request will be aborted in favor of the new XHR request. However, even though the former request is being aborted it is still being sent through all the 'success' event handlers and eventually to the 'onXhr' method, without a clear indicator that the request has been aborted. The only reliable way of handling these edge cases is by only validating the response of a completed request.